% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop_in_group.R
\name{prop_in_group}
\alias{prop_in_group}
\title{<Function for Proportions>}
\usage{
prop_in_group(input_df, value_col, breakdowns_vector, group_by_col = NULL)
}
\arguments{
\item{input_df}{an input data frame}

\item{value_col}{values}

\item{breakdowns_vector}{descrip}

\item{group_by_col}{desc}
}
\description{
<full descrip>
}
